+++ src/kaczmarz/_variants.py
@@ -6,20 +6,7 @@ import numpy as np
 from scipy import sparse
 
 import kaczmarz
-
-
-def scale_rows(A, v):
-    return sparse.spdiags(v, 0, len(v), len(v)) @ A
-
-
-def scale_cols(A, v):
-    return A @ sparse.spdiags(v, 0, len(v), len(v))
-
-
-def square(A):
-    if sparse.issparse(A):
-        return A.power(2)
-    return A ** 2
+from ._utils import scale_rows, scale_cols, square
 
 
 class Cyclic(kaczmarz.Base):
